pyproject.toml
[project]
name = "rag-gemini-pipeline"
version = "0.1.0"
description = "A Retrieval-Augmented Generation pipeline using local embeddings (SentenceTransformers), FAISS vector store, and Google Gemini Flash for query answering."
authors = [
    { name="Chiranjeevi Sai Moka", email="your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "langchain>=0.2.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=2.2.2",
    "tiktoken>=0.7.0",
    "pypdf>=3.17.1",
    "python-docx>=1.1.0",
    "google-generativeai>=0.7.2",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
